Refactor functions into separate files for modularity
